\name{bowa-package}
\alias{bowa-package}
\alias{bowa}
\docType{package}
\title{BOWA: Bereken Onzekerheid WAteropgave}
\description{Dit pakket is een pakket met functies voor BOWA, een tool om de
onzekerheid in de wateropgave te simuleren. Alhoewel het mogelijk is, is het
niet de bedoeling om de functies in dit pakket direct te gebruiken. Gebruik
liever de bijbehorende ArcGIS toolbox.}
\details{
\tabular{ll}{
Package: \tab bowa\cr
Type: \tab Package\cr
Version: \tab 0.3-2\cr
Date: \tab 2012-08-17\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maarten-Jan Kallen \email{m.j.kallen@hkv.nl}
}
\references{
\emph{Functioneel ontwerp van BOWA (versie 4)}, notitie PR2100.10, HKV Lijn in water,
8 juli 2011.
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
data(werknormen)
}
% vim: ft=tex tw=76
